Add BaseCommand
This class will act as a base class for the commands made without
the full stack framework.
